+++ pennylane/ops/functions/single_qubit_zyz_angles.py
@@ -16,8 +16,24 @@
 
 from functools import singledispatch
 
+import numpy as np
+
 from pennylane.estimator.ops.op_math.symbolic import Adjoint
 from pennylane.operation import Operator
+from pennylane.ops.qubit import (
+    RX,
+    RY,
+    RZ,
+    SX,
+    Hadamard,
+    PauliX,
+    PauliY,
+    PauliZ,
+    PhaseShift,
+    Rot,
+    S,
+    T,
+)
 from pennylane.typing import TensorLike
 
 
@@ -25,6 +41,77 @@ from pennylane.typing import TensorLike
 def single_qubit_zyz_angles(op: Operator) -> tuple[TensorLike, TensorLike, TensorLike, TensorLike]:
     """Returns the rotation angles for the ZYZ decomposition of this operator."""
     raise NotImplementedError
+
+
+@single_qubit_zyz_angles.register
+def _h_rot_angles(op: Hadamard):  # pylint: disable=unused-argument
+    # H = RZ(\pi) RY(\pi/2) RZ(0)
+    return (np.pi, np.pi / 2, 0.0, 0.0)
+
+
+@single_qubit_zyz_angles.register
+def _x_rot_angles(op: PauliX):  # pylint: disable=unused-argument
+    # X = RZ(-\pi/2) RY(\pi) RZ(\pi/2)
+    return (np.pi / 2, np.pi, -np.pi / 2, 0.0)
+
+
+@single_qubit_zyz_angles.register
+def _y_rot_angles(op: PauliY):  # pylint: disable=unused-argument
+    # Y = RZ(0) RY(\pi) RZ(0)
+    return (0.0, np.pi, 0.0, 0.0)
+
+
+@single_qubit_zyz_angles.register
+def _z_rot_angles(op: PauliZ):  # pylint: disable=unused-argument
+    # Z = RZ(\pi) RY(0) RZ(0)
+    return (np.pi, 0.0, 0.0, 0.0)
+
+
+@single_qubit_zyz_angles.register
+def _s_rot_angles(op: S):  # pylint: disable=unused-argument
+    # S = RZ(\pi/2) RY(0) RZ(0)
+    return (np.pi / 2, 0.0, 0.0, 0.0)
+
+
+@single_qubit_zyz_angles.register
+def _t_rot_angles(op: T):  # pylint: disable=unused-argument
+    # T = RZ(\pi/4) RY(0) RZ(0)
+    return (np.pi / 4, 0.0, 0.0, 0.0)
+
+
+@single_qubit_zyz_angles.register
+def _sx_rot_angles(op: SX):  # pylint: disable=unused-argument
+    # SX = RZ(-\pi/2) RY(\pi/2) RZ(\pi/2)
+    return (np.pi / 2, np.pi / 2, -np.pi / 2, 0.0)
+
+
+@single_qubit_zyz_angles.register
+def _(op: RX):
+    # RX(\theta) = RZ(-\pi/2) RY(\theta) RZ(\pi/2)
+    return (np.pi / 2, op.data[0], -np.pi / 2, 0.0)
+
+
+@single_qubit_zyz_angles.register
+def _(op: RY):
+    # RY(\theta) = RZ(0) RY(\theta) RZ(0)
+    return (0.0, op.data[0], 0.0, 0.0)
+
+
+@single_qubit_zyz_angles.register
+def _(op: RZ):
+    # RZ(\theta) = RZ(\theta) RY(0) RZ(0)
+    return (op.data[0], 0.0, 0.0, 0.0)
+
+
+@single_qubit_zyz_angles.register
+def _(op: PhaseShift):
+    # PhaseShift(\theta) = RZ(\theta) RY(0) RZ(0)
+    return (op.data[0], 0.0, 0.0, 0.0)
+
+
+@single_qubit_zyz_angles.register
+def _(op: Rot):
+    return tuple(op.data) + (0.0,)
 
 
 @single_qubit_zyz_angles.register
